Generate pyi Stub file for python bindings in POC (#3)
<!-- Edit And Paste Into Pull Request Title
Issue-<Github Issue Number>: PR Title
-->

### Changes
- Adds tooling for python stub generation to support python type hinting
- Adds scripting to verify that generated stub file stays up to date
with lib.rs

### Context For Reviewers
Python supports type hinting which allows libraries to specify what
types a function should accept and return. But since we are generating a
C library with PYO3 we need to add on the type hints separately.

I found an extension library that is fairly new for pyo3 that generates
a pyi file with the type hints defined and integrated that into the
project. I also setup the gh action to fail if the stub generation is
run before the PR can be merged.