Add YAML Config Source to Core POC (#7)
<!-- Edit And Paste Into Pull Request Title
Issue-<Github Issue Number>: PR Title
-->

<!--NOTE THE GITHUB ISSUES THAT THIS PR CLOSES BELOW-->
closes #

### Changes
<!-- DESCRIBE THE CHANGES YOU ARE MAKING -->
- Adds YAML ConfigSource to core POC
- adds yaml-rust2 as dependency to core
- Adds Yaml ConfigSource to ConfigBuilder

### Context For Reviewers
<!-- WHY ARE YOU MAKING THIS CHANGE? -->

YAML is a great format for configuration files and one that I think this
project should try to support the best. This PR adds support for YAML as
a config source and allows it to be specified in the ConfigBuilder to
ensure that yaml configuration is used just like any other source.

there are some details to work out more with this such as supporting
lists. But that will be done as a follow up